# public/wapd-api-doc.json
{
  "swagger": "2.0",
  "info": {
    "description": "Wapd api document for front-end developer",
    "version": "0.0.1",
    "title": "Wapd",
    "contact": {
      "email": "uu9924079@gmail.com"
    }
  },
  "host": "localhost:3000",
  "schemes": ["http", "https"],
  "securityDefinitions": {
    "Authorization": {
      "type": "apiKey",
      "in": "header",
      "name": "Authorization"
    }
  },
...
  "paths": {
...
    "/api/users/edit": {
      "get": {
        "tags": ["products"],
        "summary": "Return user data",
        "description": "",
        "operationId": "userController.getUserEdit",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "Authorization",
            "type": "string",
            "in": "header",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "object",
              "properties": {
                "user": {
                  "properties": {
                    "id": {
                      "type": "integer"
                    },
                    "email": {
                      "type": "string"
                    },
                    "name": {
                      "type": "string"
                    },
                    "phone": {
                      "type": "string"
                    },
                    "address": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["users"],
        "summary": "Return result",
        "description": "",
        "operationId": "userController.putUser",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "Authorization",
            "type": "string",
            "in": "header",
            "required": true
          },
          {
            "in": "body",
            "name": "body",
            "description": "personal data you want to change",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string"
                },
                "email": {
                  "type": "string",
                  "required": "true"
                },
                "birthday": {
                  "type": "string"
                },
                "address": {
                  "type": "string"
                },
                "phone": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "object"
            }
          }
        }
      }
    },
    "/api/users/password_change": {
      "post": {
        "tags": ["users"],
        "summary": "Change user's password",
        "description": "",
        "operationId": "userController.postPasswordChange",
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Your personal data",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "usedPassword": {
                  "type": "string"
                },
                "newPassword": {
                  "type": "string"
                },
                "passwordCheck": {
                  "type": "string"
                }
              }
            }
          },
          {
            "name": "Authorization",
            "type": "string",
            "in": "header",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object"
            }
          }
        }
      }
    },
    "/api/users/password_forget": {
      "post": {
        "tags": ["users"],
        "summary": "Send an token to reset user's password",
        "description": "",
        "operationId": "userController.postPasswordForget",
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Your email",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object"
            }
          }
        }
      }
    },
    "/api/users/password_reset/{token_id}/{token}": {
      "get": {
        "tags": ["users"],
        "summary": "Reset password",
        "description": "",
        "operationId": "userController.getPasswordReset",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "token_id",
            "in": "path",
            "description": "Token id",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "token",
            "in": "path",
            "description": "Token",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "object",
              "properties": {
                "product": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/api/users/password_reset": {
      "post": {
        "tags": ["users"],
        "summary": "Reset user's password",
        "description": "",
        "operationId": "userController.postPasswordReset",
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Your email",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "token": {
                  "type": "string"
                },
                "userId": {
                  "type": "integer"
                },
                "password": {
                  "type": "string"
                },
                "passwordCheck": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object"
            }
          }
        }
      }
    },
...
    "/api/users/wishlist": {
      "get": {
        "tags": ["users"],
        "summary": "Display user's wishlist",
        "description": "returns an array of objects",
        "operationId": "userController.getUserWishlist",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "Authorization",
            "type": "string",
            "in": "header",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "object",
              "properties": {
                "products": {
                  "type": "array",
                  "example": "[{product object}, {product object}]"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/users/orders": {
      "get": {
        "tags": ["users"],
        "summary": "Display user's orders",
        "description": "returns an array of objects",
        "operationId": "userController.getUserOrders",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "Authorization",
            "type": "string",
            "in": "header",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "object",
              "properties": {
                "products": {
                  "type": "array",
                  "example": "[{order object}, {order object}]"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["orders"],
        "summary": "Adding product to order from cart",
        "description": "",
        "operationId": "productController.addWishlist",
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Coupon you want to use",
            "schema": {
              "type": "object",
              "properties": {
                "coupon": {
                  "type": "string"
                }
              }
            }
          },
          {
            "name": "Authorization",
            "type": "string",
            "in": "header",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object"
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/users/orders/{order_id}": {
      "get": {
        "tags": ["users"],
        "summary": "Display an order by id",
        "description": "",
        "operationId": "userController.getUserOrder",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "order_id",
            "in": "path",
            "description": "ID of order to return",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "Authorization",
            "type": "string",
            "in": "header",
            "required": true
          }
        ],
        "security": [
          {
            "Authorization": []
          }
        ],
        "responses": {
          "200": {
            "description": "Returns an object",
            "schema": {
              "type": "object",
              "properties": {
                "orders": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "integer"
                    },
                    "total_price": {
                      "type": "integer"
                    },
                    "shipping_status": {
                      "type": "string"
                    },
                    "shipping_method": {
                      "type": "string"
                    },
                    "payment_status": {
                      "type": "string"
                    },
                    "phone": {
                      "type": "string"
                    },
                    "email": {
                      "type": "string"
                    },
                    "payment_method": {
                      "type": "string"
                    },
                    "address": {
                      "type": "string"
                    },
                    "receiver_name": {
                      "type": "string"
                    },
                    "comment": {
                      "type": "string"
                    },
                    "sn": {
                      "type": "string"
                    },
                    "UserId": {
                      "type": "integer"
                    },
                    "CouponId": {
                      "type": "integer"
                    },
                    "createdAt": {
                      "type": "string"
                    },
                    "updatedAt": {
                      "type": "string"
                    },
                    "Coupon": {
                      "type": "string"
                    },
                    "items": {
                      "type": "array",
                      "example": "[{item Object}, {item Object}]"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
    "/api/admins/users/{id}/orders": {
      "get": {
        "tags": ["admins"],
        "summary": "Display user's orders by id",
        "description": "",
        "operationId": "adminController.getUserOrders",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of order to return",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "Authorization",
            "type": "string",
            "in": "header",
            "required": true
          }
        ],
        "security": [
          {
            "Authorization": []
          }
        ],
        "responses": {
          "200": {
            "description": "Returns an object",
            "schema": {
              "type": "object",
              "properties": {
                "users": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "integer"
                    },
                    "email": {
                      "type": "string"
                    },
                    "name": {
                      "type": "string"
                    },
                    "phone": {
                      "type": "string"
                    },
                    "address": {
                      "type": "string"
                    },
                    "orders": {
                      "type": "array",
                      "example": "[{orderItem Object}, {orderItem Object}]"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
  }
}